Refactor (#12)
* Break up structure with directories
* Split out Option
* Option card areas are themselves toggles
